+++ pkg/internal/procs/proclang_test.go
@@ -4,13 +4,11 @@
 package procs
 
 import (
-	"encoding/binary"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
 	"go.opentelemetry.io/obi/pkg/appolly/app/svc"
-	"go.opentelemetry.io/obi/pkg/internal/fastelf"
 )
 
 func TestHighCertaintyModuleDetection(t *testing.T) {
@@ -79,33 +77,4 @@ func TestLastResortDetection(t *testing.T) {
 	assert.Equal(t, svc.InstrumentableGeneric, instrumentableLastResort("libstdc++"))
 	assert.Equal(t, svc.InstrumentableGeneric, instrumentableLastResort("libc++"))
 	assert.Equal(t, svc.InstrumentableGeneric, instrumentableLastResort("/usr/lib/libsomething.so"))
-}
-
-func TestMatchExeSymbols_InvalidStringOffset(t *testing.T) {
-	const symSize = 24
-
-	data := make([]byte, symSize+4)
-	binary.LittleEndian.PutUint32(data[0:4], 128)
-	data[4] = 0x02
-	binary.LittleEndian.PutUint64(data[8:16], 1)
-	binary.LittleEndian.PutUint64(data[16:24], 1)
-	copy(data[symSize:], []byte("x\x00"))
-
-	ctx := &fastelf.ElfContext{
-		Data: data,
-		Sections: []*fastelf.Elf64_Shdr{
-			{
-				Type:    fastelf.SHT_SYMTAB,
-				Link:    1,
-				Offset:  0,
-				Size:    symSize,
-				Entsize: symSize,
-			},
-			{
-				Offset: symSize,
-			},
-		},
-	}
-
-	assert.Equal(t, svc.InstrumentableGeneric, matchExeSymbols(ctx))
 }
